+++ src/realm/table.hpp
@@ -1396,13 +1396,13 @@ inline void Table::bump_version(bool bump_global) const REALM_NOEXCEPT
 inline void Table::remove(size_t row_ndx)
 {
     bool is_move_last_over = false;
-    erase_row(row_ndx, is_move_last_over); // Throws;
+    erase_row(row_ndx, is_move_last_over); // Throws
 }
 
 inline void Table::move_last_over(size_t row_ndx)
 {
     bool is_move_last_over = true;
-    erase_row(row_ndx, is_move_last_over); // Throws;
+    erase_row(row_ndx, is_move_last_over); // Throws
 }
 
 inline void Table::remove_last()
@@ -2220,6 +2220,11 @@ public:
     static bool is_cross_table_link_target(const Table& table)
     {
         return table.is_cross_table_link_target();
+    }
+
+    static Group* get_parent_group(const Table& table) REALM_NOEXCEPT
+    {
+        return table.get_parent_group();
     }
 
     static Replication* get_repl(Table& table) REALM_NOEXCEPT