+++ release_notes.md
@@ -22,6 +22,8 @@
 
 * Added `Descriptor::get_link_target()`, for completeness.
 * Added extra `allow_file_format_upgrade` argument to `SharedGroup::open()`.
+* Modifying `Descriptor` methods now throw `LogicError` when appropriate (rather
+  than asserting).
 
 -----------